feat: layout changes
