+++ src/app/globals.css
@@ -1,3 +1,11 @@
 body {
   margin: 0;
 }
+
+.section {
+  margin: 0 20px;
+}
+
+.section-bg-color {
+  padding: 0 20px;
+}
