+++ src/app/not-found.tsx
@@ -1,13 +1,8 @@
-import Link from "next/link";
-
 export default async function NotFound() {
   return (
-    <div>
-      <h2>404 - Page Not Found</h2>
-      <p>Could not find requested page</p>
-      <p>
-        Visit <Link href="/">homepage</Link>
-      </p>
+    <div className="section">
+      <h2>Page Not Found</h2>
+      <p>Oops, could not find requested page</p>
     </div>
   );
 }
