+++ src/components/Header/Header.module.css
@@ -8,7 +8,6 @@
   background-color: #703809;
   justify-content: space-between;
   align-items: center;
-  padding: 0 20px;
   box-sizing: border-box;
   z-index: 90;
 }
