+++ src/components/Layout/Layout.module.css
@@ -1,0 +1,4 @@
+.TopSpacing {
+  padding-top: 70px;
+  width: 100vw;
+}
